attempt fix visual latch drift
