+++ Content.Shared/_Starlight/Actions/Components/LatchComponent.cs
@@ -65,16 +65,6 @@ public sealed partial class LatchComponent : Component
     public float DriftBreakTolerance = 0.5f;
 
     /// <summary>
-    /// Hard ceiling on how far the grace-period recovery pull (see Update()) is allowed
-    /// to move the latcher to reach the target. Meant only to smooth minor jitter right
-    /// after a latch starts; if the target is farther than this, something else moved
-    /// it (or the target was wrong to begin with) and the latch should just break instead
-    /// of dragging the latcher along, however far, to reach it.
-    /// </summary>
-    [DataField]
-    public float MaxDriftPullDistance = 3f;
-
-    /// <summary>
     /// Cap on the physics joint's max length. Matches baseline unarmed melee
     /// range (1.5), not DriftBreakRange, so the target can always punch back.
     /// </summary>
